split top-level headers
